What is the difference between an iterator and a generator
First, recall what an iterator is it's any object that has __iter__() and __next() methods. The first returns the iterator objects itself and is used in for and in statements. The second method returns the next value in the iteration --if there is no more items it raises the StopIteration exception. Here is an iterator that returns numbers that an increment by a random amount between 0 and 1, stopping at a limit specified in the constructor.