Add files via upload
A generator is an easy way to create iterators. As a concrete example, here is the same iterator expressed as a generator. Note the use of the keyword yield. A generator uses the function call stack to implicity store the state of the iterator __this can simplify the writing of an iterator compared to writing the same iterator as an explicit class; it also helps readability.