Add files via upload
Writing a program to count the number of bits that are set to 1 in a non-negative integer is a good way to get up to speed with primitive types. This program tests bits one-at-a-time starting with the least significant bit. It illustrates shifting and masking and also shows how to avoid hard-coding the size of the integer word.