Computing the Parity of a word
The parity of a binary ord is 1 if the number of 1s in the word is odd; otherwise, it is 0. For example, the parity of 1011 is 1, and the parity of 10001000 is 0. Parity checks are used to detect single-bit errors in data storage and communication. It is fairly straightforward to write code that computes the parity of a single 64-bit word. 
This is a solution code that how I did the computation of the parity of a very large number of 64-bit words.
Using the time complexity is O(n) where n is the word size, describing the several algorithms for parity computation that are superior to the brute-force algorithm