SWAP BITS
There is a number of ways in which bit manipulations can be accelerated. For like the expression  x & (x-1) --clears the lowest set bit in x, and x& ~(x-1) extracts the lowest set bit of x. A 64-bit integer can be viewed as an array of 64 bits, with the bit at index 0 correspondings to the least significant bit (LSB), and the bit at index 63 correspondings to the most significant bit (MSB) Implement code that takes as input a 64-bit integer and swaps the bits at indices i and j. 
When is the swap necessary?
ANS:  If we were swapping objects stored in an array, However, since a bit can only take two values, we can do a little better. Specifically, we first test if the bits to be swapped differ. If they do not, the swap does not change the integer.