Find A Closest Integer with the same weight
A program which takes as input a non-negative integer x and returns a number y which is not equal to x, but has the same weight as x and their difference |y -x| is as small as possible.

Solution Approach:  Might try all the integers x-1 , x+1, x-2, x+2..... stopping as soon as we encounter one with the same weight at x. This performs very poorly on some inputs one way to see this is to consider the case where x=2^3 =8 The only numbers with a weight of 1 are powers of 2. Thus the algorithm will try the following sequence 7,9,6,10,5,11,4, stopping at 4